move common code and create lib
