+++ lib/query.js
@@ -1,3 +1,5 @@
+import _ from 'lodash';
+
 const mmolL = 'mmol/L';
 const mgdL = 'mg/dL';
