+++ lib/userDataHandler.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import queryString from 'query-string';
 import dataTools from '@tidepool/data-tools';
-import { config, log, buildHeaders, createCounter } from './app';
+import { buildHeaders, createCounter, exportTimeout } from './utils';
 
 const dataStatusCount = createCounter(
   'tidepool_export_status_count',
@@ -9,12 +9,16 @@ const dataStatusCount = createCounter(
   ['status_code', 'export_format']
 );
 
-export function userDataExportHandler() {
+const log = logMaker('userDataHandler.js', {
+  level: process.env.DEBUG_LEVEL || 'info',
+});
+
+export function userDataHandler() {
   return async (req, res) => {
     // Set the timeout for the request. Make it 10 seconds longer than
     // our configured timeout to give the service time to cancel the API data
     // request, and close the outgoing data stream cleanly.
-    req.setTimeout(config.exportTimeout + 10000);
+    req.setTimeout(exportTimeout + 10000);
 
     const queryData = [];
 
@@ -80,8 +84,8 @@ export function userDataExportHandler() {
       // Create a timeout timer that will let us cancel the incoming request gracefully if
       // it's taking too long to fulfil.
       const timer = setTimeout(() => {
-        res.emit('timeout', config.exportTimeout);
-      }, config.exportTimeout);
+        res.emit('timeout', exportTimeout);
+      }, exportTimeout);
 
       // Wait for the stream to complete, by wrapping the stream completion events in a Promise.
       try {
@@ -90,7 +94,10 @@ export function userDataExportHandler() {
           dataResponse.data.on('error', (err) => reject(err));
           res.on('error', (err) => reject(err));
           res.on('timeout', async () => {
-            dataStatusCount.inc({ status_code: 408, export_format: exportFormat });
+            dataStatusCount.inc({
+              status_code: 408,
+              export_format: exportFormat,
+            });
             reject(
               new Error(
                 'Data export request took too long to complete. Cancelling the request.'
