+++ lib/userReportHandler.js
@@ -1,7 +1,6 @@
-import { config, createCounter } from './app';
-
 import { DataUtil } from '@tidepool/viz/dist/data';
-import { validateQueryParams } from './report_queries/query';
+import { validateQueryParams } from './query';
+import { buildHeaders, createCounter, exportTimeout } from './utils';
 
 const reportStatusCount = createCounter(
   'tidepool_export_report_status_count',
@@ -11,12 +10,16 @@ const reportStatusCount = createCounter(
 
 var dataUtil = new DataUtil();
 
-export function userReportExportHandler() {
+const log = logMaker('userReportHandler.js', {
+  level: process.env.DEBUG_LEVEL || 'info',
+});
+
+export function userReportHandler() {
   return async (req, res) => {
     // Set the timeout for the request. Make it 10 seconds longer than
     // our configured timeout to give the service time to cancel the API data
     // request, and close the outgoing data stream cleanly.
-    req.setTimeout(config.exportTimeout + 10000);
+    req.setTimeout(exportTimeout + 10000);
 
     if (!req.body) {
       reportStatusCount.inc({ status_code: 400, report_params: {} });
@@ -53,14 +56,14 @@ export function userReportExportHandler() {
       );
       log.debug(`Downloading data for User ${req.params.userid}...`);
 
-      var { reports } =  req.body;
+      var { reports } = req.body;
       dataUtil.addData(dataResponse.data, userId);
 
       // Create a timeout timer that will let us cancel the incoming request gracefully if
       // it's taking too long to fulfil.
       const timer = setTimeout(() => {
-        res.emit('timeout', config.exportTimeout);
-      }, config.exportTimeout);
+        res.emit('timeout', exportTimeout);
+      }, exportTimeout);
 
       // Wait for the stream to complete, by wrapping the stream completion events in a Promise.
       try {
